a real bundle of changes

+++ leveln/leveln_writer_test.go
@@ -37,7 +37,7 @@ func BenchmarkLevelNAppend(b *testing.B) {
 				ln.Init(keys, values)
 
 				for j := 0; j < n; j++ {
-					key = testhelp.KeyFrom(uint64(j)/32, 0, uint32(j))
+					key = testhelp.KeyFrom(uint32(j)/32, 0, uint32(j))
 					_ = ln.Append(key, value[0:256+rng.Uint32()%256])
 				}
 				_ = ln.Finish()
